+++ local-development.php
@@ -14,7 +14,7 @@
  * Author:            Andy Fragen
  * Author URI:        http://thefragens.com/
  * Description:       Places development notice for plugins or themes that are in local development. Prevents updating of selected plugins and themes. Allows for using a local file server and bypassing the WordPress 5.2 WSOD Shutdown Handler.
- * Version:           2.4.1
+ * Version:           2.4.1.1
  * Domain Path:       /languages
  * Text Domain:       local-development
  * License:           GNU General Public License v2
@@ -22,7 +22,7 @@
  * Network:           true
  * GitHub Plugin URI: https://github.com/afragen/local-development
  * Requires PHP:      5.6
- * Requires WP:       4.0
+ * Requires at least: 4.0
  */
 
 namespace Fragen\Local_Development;