+++ CHANGES.md
@@ -2,7 +2,8 @@
 * update `Requires at least` in plugin file
 * add VCS checkouts automatically, thanks @Raruto
 * add Git Host icons as default setting for plugins and themes
-* automatically add localhost admin bar coloring when on localhost, default is on
+* add localhost admin bar coloring when on localhost, default is on
+* some now settings only display when running in localhost
 * add repositories that are added from [GitHub Update Additions](https://github.com/afragen/github-updater-additions)
 
 #### 2.4.1 / 2019-05-09