+++ CHANGES.md
@@ -4,7 +4,7 @@
 * add Git Host icons as default setting for plugins and themes
 * add localhost admin bar coloring when on localhost, default is on
 * some now settings only display when running in localhost
-* add repositories that are added from [GitHub Update Additions](https://github.com/afragen/github-updater-additions)
+* add repositories that are added from [GitHub Updater Additions](https://github.com/afragen/github-updater-additions)
 
 #### 2.4.1 / 2019-05-09
 * a11y updates for settings tabs